projeto_criptografia/requirements.txt
streamlit
pycryptodome
cryptography